# docs/swagger.yaml
basePath: /api/v1
definitions:
  realNumberDNOClone_internal_models.AddDNORequest:
    properties:
      channel:
        type: string
      numberType:
        type: string
      phoneNumber:
        type: string
      reason:
        type: string
    type: object
  realNumberDNOClone_internal_models.AnalyticsSummary:
    properties:
      activeNumbers:
        type: integer
      byChannel:
        additionalProperties:
          type: integer
        type: object
      byDataset:
        additionalProperties:
          type: integer
        type: object
      byNumberType:
        additionalProperties:
          type: integer
        type: object
      hitRate24h:
        type: number
      queriesByHour:
        items:
          $ref: '#/definitions/realNumberDNOClone_internal_models.HourlyCount'
        type: array
      recentAdditions:
        type: integer
      recentRemovals:
        type: integer
      totalDnoNumbers:
        type: integer
      totalQueries24h:
        type: integer
    type: object
  realNumberDNOClone_internal_models.AuditLog:
    properties:
      action:
        type: string
      createdAt:
        type: string
      details:
        type: string
      entityId:
        type: integer
      entityType:
        type: string
      id:
        type: integer
      orgId:
        type: integer
      userId:
        type: integer
    type: object
  realNumberDNOClone_internal_models.AuditLogPage:
    properties:
      data:
        items:
          $ref: '#/definitions/realNumberDNOClone_internal_models.AuditLog'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  realNumberDNOClone_internal_models.BulkJob:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      errorCount:
        type: integer
      fileName:
        type: string
      id:
        type: integer
      jobType:
        type: string
      orgId:
        type: integer
      processedRecords:
        type: integer
      resultSummary:
        type: string
      status:
        type: string
      successCount:
        type: integer
      totalRecords:
        type: integer
      userId:
        type: integer
    type: object
...
  realNumberDNOClone_internal_models.DNONumber:
    properties:
      addedByOrgId:
        type: integer
      addedByUserId:
        type: integer
      channel:
        type: string
      createdAt:
        type: string
      dataset:
        type: string
      id:
        type: integer
      investigationId:
        type: string
      numberType:
        type: string
      phoneNumber:
        type: string
      reason:
        type: string
      status:
        type: string
      threatCategory:
        type: string
      updatedAt:
        type: string
    type: object
  realNumberDNOClone_internal_models.DNONumberPage:
    properties:
      data:
        items:
          $ref: '#/definitions/realNumberDNOClone_internal_models.DNONumber'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  realNumberDNOClone_internal_models.HourlyCount:
    properties:
      count:
        type: integer
      hour:
        type: string
    type: object
...
  realNumberDNOClone_internal_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  realNumberDNOClone_internal_models.LoginResponse:
    properties:
      refreshToken:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/realNumberDNOClone_internal_models.User'
    type: object
...
  realNumberDNOClone_internal_models.User:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      orgId:
        type: integer
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: FakeNumber DNO
    url: https://github.com/drewkosta/realNumberDNOClone
  description: |-
    Do Not Originate database API for preventing illegal robocalls and caller ID spoofing.
    Carriers query the DNO database in real-time to block spoofed calls.
  title: FakeNumber DNO API
  version: "1.0"
paths:
...
  /analytics:
    get:
      description: Retrieve DNO analytics summary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.AnalyticsSummary'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get analytics
      tags:
      - Analytics
...
  /audit-log:
    get:
      description: Retrieve a paginated audit log of DNO operations
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.AuditLogPage'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log
      tags:
      - Analytics
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, returns JWT access and refresh
        tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/realNumberDNOClone_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          properties:
            refreshToken:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh token
      tags:
      - Auth
...
  /dno/bulk-job:
    get:
      description: Retrieve the status of a bulk upload job
      parameters:
      - description: Bulk job ID
        in: query
        name: jobId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.BulkJob'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get bulk job status
      tags:
      - Bulk Operations
  /dno/bulk-upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a CSV file of phone numbers for bulk addition to the DNO
        list
      parameters:
      - description: CSV file of phone numbers
        in: formData
        name: file
        required: true
        type: file
      - description: Channel (defaults to voice)
        in: formData
        name: channel
        type: string
      - description: Number type (defaults to local)
        in: formData
        name: numberType
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            properties:
              jobId:
                type: integer
              message:
                type: string
              status:
                type: string
              totalRecords:
                type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk upload DNO numbers
      tags:
      - Bulk Operations
  /dno/export:
    get:
      description: Download all DNO numbers as a CSV file
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Export DNO numbers as CSV
      tags:
      - Bulk Operations
  /dno/numbers:
    delete:
      description: Remove a phone number from the Do-Not-Originate list
      parameters:
      - description: Phone number to remove
        in: query
        name: phoneNumber
        required: true
        type: string
      - description: Channel (defaults to voice)
        in: query
        name: channel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a DNO number
      tags:
      - DNO Management
    get:
      description: Retrieve a paginated list of DNO numbers with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Filter by dataset
        in: query
        name: dataset
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by channel
        in: query
        name: channel
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.DNONumberPage'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List DNO numbers
      tags:
      - DNO Management
    post:
      consumes:
      - application/json
      description: Register a phone number on the Do-Not-Originate list
      parameters:
      - description: Number to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/realNumberDNOClone_internal_models.AddDNORequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/realNumberDNOClone_internal_models.DNONumber'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a DNO number
      tags:
      - DNO Management
...
securityDefinitions:
  APIKeyAuth:
    description: Organization API key for machine-to-machine access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT Bearer token (prefix with "Bearer ")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"